@@ -15,11 +15,11 @@
     "test:e2e:ui": "playwright test --ui tests/e2e/basic.spec.ts",
     "test:visual": "tsx scripts/visual-test.ts",
     "check:links": "playwright test tests/e2e/external-links.spec.ts",
-    "generate:rss": "tsx scripts/generate-rss.js",
+    "generate:rss": "tsx scripts/generate-rss.cjs",
     "generate:sitemap": "next-sitemap",
-    "check:sitemap": "node scripts/check-sitemap.js",
+    "check:sitemap": "node scripts/check-sitemap.cjs",
     "validate": "npm run lint && npm run type-check && npm run check:sitemap && npm run test",
-    "generate:slideshow": "node scripts/generate-slideshow.js",
+    "generate:slideshow": "node scripts/generate-slideshow.cjs",
     "analyze": "ANALYZE=true npm run build",
     "dev:clean": "rm -rf .next && npm run dev",
     "knip": "knip",
@@ -28,13 +28,12 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "commonjs",
+  "type": "module",
   "dependencies": {
-    "@eslint/eslintrc": "^3.3.1",
     "@heroicons/react": "^2.2.0",
     "@mdx-js/loader": "^3.1.0",
     "@mdx-js/react": "^3.1.0",
-    "@next/mdx": "^15.5.2",
+    "@next/mdx": "^16.1.1",
     "@tailwindcss/typography": "^0.5.16",
     "@types/reveal.js": "^5.2.0",
     "@vercel/analytics": "^1.5.0",
@@ -42,7 +41,7 @@
     "date-fns": "^4.1.0",
     "feed": "^5.1.0",
     "gray-matter": "^4.0.3",
-    "next": "^15.5.2",
+    "next": "^16.1.1",
     "next-sitemap": "^4.2.3",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
@@ -58,6 +57,7 @@
   },
   "devDependencies": {
     "@jest/globals": "^30.2.0",
+    "@next/eslint-plugin-next": "^16.1.1",
     "@playwright/test": "^1.55.0",
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/react": "^16.3.0",
@@ -69,15 +69,18 @@
     "@types/testing-library__jest-dom": "^5.14.9",
     "autoprefixer": "^10.4.21",
     "eslint": "^9.34.0",
-    "eslint-config-next": "^15.5.4",
+    "eslint-config-next": "^16.1.1",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.1",
     "jest": "^30.1.1",
     "jest-environment-jsdom": "^30.1.1",
     "knip": "^5.64.1",
     "postcss": "^8.5.6",
     "tsx": "^4.20.5",
-    "typescript": "^5.9.2"
+    "typescript": "^5.9.2",
+    "typescript-eslint": "^8.52.0"
   },
   "optionalDependencies": {
-    "@next/swc-darwin-arm64": "^15.5.2"
+    "@next/swc-darwin-arm64": "^16.1.1"
   }
 }
